+++ packages/web/CHANGELOG.md
@@ -5,4 +5,11 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @rocketseat/houston-contracts@0.1.7
+
+## null
+
+### Patch Changes
+
+- Updated dependencies
   - @rocketseat/houston-contracts@0.1.6